Create code_photo
The code opens a txt file and pulls a link to the photo from it, the link does not give the direct address of the photo. For this reason it uses the driver for google chrome and opens the photo in it. It has a built in timer, if the link doesn't open within 10 sec it will skip the link. After it opens the photo, it copies it and saves the photo to the folder where the script is running. Once the photo is saved, the browser closes. The actions continue as long as there are links in the txt files.